(PC-26667) refactor(api): Move refreshAccessTokenWithRetriesOnError
